Reimplemented DSE loading completely
* Rewrote the loading of DSE files completely. The old code was a spaghetti mess and extremely redundant.
* Rewrote dse to soundfont conversion handling to be simpler.
- Sample baking currently doesn't work at all though.

+++ ppmdu_gfxcrunch/src/gfxcrunch.cpp
@@ -839,7 +839,7 @@ namespace gfx_util
                 ++i;
             }
 
-            updtProgress = std::async( std::launch::async, PrintProgressLoop, std::ref(completed), mysprites.size(), std::ref(shouldUpdtProgress) );
+            updtProgress = std::async( std::launch::async, PrintProgressLoop, std::ref(completed), static_cast<uint32_t>(mysprites.size()), std::ref(shouldUpdtProgress) );
             
             taskmanager.Execute();
             taskmanager.BlockUntilTaskQueueEmpty();
